move retroarch memory map setup to device.c
